Tests for aws and azure are separated
