+++ phpunit.xml.dist
@@ -16,6 +16,12 @@
 		<directory>tests/Common</directory>
 		<directory>tests/Options</directory>
 		<directory>tests/S3Uploader</directory>
+		<file>tests/File/CommonFilesTests.php</file>
+		<file>tests/File/AwsFileTest.php</file>
+	</testsuite>
+
+	<testsuite name="files-azure">
+		<file>tests/File/AzureFileTest.php</file>
 	</testsuite>
 
 	<testsuite name="backend-redshift-part-1">
